# docs/design/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "autohunt API",
    "description": "autohunt 系统对外 API 契约（技术设计 v1.3 §3）。\n\n通用约定：JSON；错误统一信封 `{\"error\": {\"code\", \"message\", \"details?\"}}`；时间一律 RFC3339（UTC 存储、前端本地化展示）；列表分页 `?cursor=&limit=`（默认 50）。\n\n鉴权（§3.1）：Web UI 使用 HttpOnly Cookie session（UISession）；外部 Agent 使用 `Authorization: Bearer ah_live_<random>`（AgentBearer）。同一套路由双鉴权，中间件按凭证类型打标 caller ∈ {ui, agent} 供状态机来源裁决（BR-11）。\n\n错误码表：UNAUTHORIZED(401)、FORBIDDEN(403)、NOT_FOUND(404)、PERMIT_REQUIRED(403)/PERMIT_INVALID(403)（submit_token 缺失或无效）、STATE_CONFLICT(409)（状态机裁决拒绝）、VALIDATION_ERROR(422)。\n\n契约 v2 增补（技设 §3.7，M3–M5 写侧，全部仅 UI session，事件详情只读除外）：简历上传/版本管理（/resumes，FR-1/2/3）、档案写（PUT /profile，FR-2）、邮箱账户绑定/解绑/状态/重授权（/email-accounts，FR-40/44）、事件确认/丢弃/修正与原文回溯（/events/{id}/confirm|discard|raw，FR-42/43，BR-2）、通知列表（/notifications，FR-32）、统计与 CSV 导出（/stats/*，FR-50/51/52，口径 §10.4）。\n\n契约 v2 修订（0.2.1）：确认流补 4 项——GET /confirmations 列表（仅 UI）、GET /confirmations/{id} 待确认变体按 caller 区分（UI 返回快照，Agent 仍仅 status）、POST /confirmations/{id}/close 手动关闭（仅 UI）、已确认响应含提交结果回写（submit_result/fail_reason/submitted_at，FR-24）；D-05 读侧三端点（/applications/{id}/history|confirmations|emails，FR-31/24/43）；提醒偏好 GET/PUT /settings/reminders（仅 UI）；GET /applications 增可选 from/to 筛选。既有 v1 冻结 19 端点语义未变动。",
    "version": "0.2.1"
  },
  "paths": {
...
    "/api/v1/applications/{application_id}/emails": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "关联邮件回溯（FR-43，D-05 邮件 Tab）【契约 v2 修订】",
        "description": "返回识别匹配到该投递所属岗位（matched_job_id = application.job_id）的全部邮箱事件，含证据区元数据（主题/发件人/收件时间，RISK-5）。按识别时间正序。",
        "operationId": "get_application_emails_api_v1_applications__application_id__emails_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailEventDetailList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          },
          {
            "AgentBearer": []
          }
        ]
      }
    },
...
    "/api/v1/email-accounts/test": {
      "post": {
        "tags": [
          "email-accounts"
        ],
        "summary": "测试 IMAP 连接（D-10「测试连接」）",
        "description": "即时验证授权码与服务器连通性，不落库；失败原因在 error 字段返回（始终 200）。",
        "operationId": "test_email_account_api_v1_email_accounts_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAccountBind"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailAccountTestResult"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/email-accounts": {
      "get": {
        "tags": [
          "email-accounts"
        ],
        "summary": "邮箱账户列表（含状态，FR-44）",
        "description": "status=auth_failed 时 UI 顶部持续警示条（AC-8）；永不包含授权码。",
        "operationId": "list_email_accounts_api_v1_email_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailAccountList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      },
      "post": {
        "tags": [
          "email-accounts"
        ],
        "summary": "绑定求职邮箱（FR-40，OP-4 IMAP 授权码）",
        "description": "绑定前先做连接验证：失败 → 422 VALIDATION_ERROR（统一信封，message 含原因），不创建账户。授权码 Fernet 加密落盘（§2.2 密钥存储口径），任何响应不回传。绑定成功即启动该账户轮询。",
        "operationId": "bind_email_account_api_v1_email_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAccountBind"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailAccountInfo"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "STATE_CONFLICT — 该邮箱已绑定",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/email-accounts/{account_id}": {
      "patch": {
        "tags": [
          "email-accounts"
        ],
        "summary": "重授权（FR-44 恢复路径）",
        "description": "授权失效后提交新授权码：验证通过 → status 恢复 active、续跑轮询（last_uid 保证不重不漏）；验证失败 → 422 VALIDATION_ERROR，status 保持 auth_failed。",
        "operationId": "reauth_email_account_api_v1_email_accounts__account_id__patch",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAccountReauth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailAccountInfo"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      },
      "delete": {
        "tags": [
          "email-accounts"
        ],
        "summary": "解绑并清除凭据（RISK-3）",
        "description": "停止轮询、删除加密凭据；历史已识别事件与日程完整保留（FR-44/AC-8）。",
        "operationId": "unbind_email_account_api_v1_email_accounts__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailAccountBind": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "imap_host": {
            "type": "string",
            "title": "Imap Host"
          },
          "port": {
            "type": "integer",
            "title": "Port",
            "default": 993
          },
          "auth_code": {
            "type": "string",
            "title": "Auth Code",
            "description": "IMAP 授权码（OP-4）；服务端 Fernet 加密落盘，任何响应不回传"
          }
        },
        "type": "object",
        "required": [
          "email",
          "imap_host",
          "auth_code"
        ],
        "title": "EmailAccountBind"
      },
      "EmailAccountInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "imap_host": {
            "type": "string",
            "title": "Imap Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "status": {
            "$ref": "#/components/schemas/EmailAccountStatus"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC3339 时间戳（UTC 存储）"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "RFC3339 时间戳（UTC 存储）"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "imap_host",
          "port",
          "status",
          "created_at"
        ],
        "title": "EmailAccountInfo",
        "description": "永不包含授权码。"
      },
      "EmailAccountList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmailAccountInfo"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EmailAccountList"
      },
      "EmailAccountReauth": {
        "properties": {
          "auth_code": {
            "type": "string",
            "title": "Auth Code",
            "description": "新授权码；验证通过则 status 恢复 active 并续跑轮询（FR-44）"
          }
        },
        "type": "object",
        "required": [
          "auth_code"
        ],
        "title": "EmailAccountReauth"
      },
      "EmailAccountStatus": {
        "type": "string",
        "enum": [
          "active",
          "auth_failed"
        ],
        "title": "EmailAccountStatus",
        "description": "FR-44：auth_failed 时暂停轮询 + UI 全局警示，历史数据保留。"
      },
      "EmailAccountTestResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "连接/认证失败原因（D-10 即时反馈）"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "EmailAccountTestResult"
      },
...
      "EmailEventDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/EmailEventType"
          },
          "event_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC3339 时间戳（UTC 存储）"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "meeting_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Link"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "matched_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/EmailEventStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "RFC3339 时间戳（UTC 存储）"
          },
          "email_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Subject"
          },
          "email_sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Sender"
          },
          "email_received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC3339 时间戳（UTC 存储）"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Received At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "created_at"
        ],
        "title": "EmailEventDetail",
        "description": "事件详情 = 列表字段 + 证据区元数据（D-07，RISK-5 可回溯）。"
      },
      "EmailEventDetailList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmailEventDetail"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EmailEventDetailList",
        "description": "邮件回溯列表（FR-43，D-05 邮件 Tab，契约 v2 修订）。"
      },
...
      "EmailEventStatus": {
        "type": "string",
        "enum": [
          "待确认",
          "已确认",
          "已丢弃"
        ],
        "title": "EmailEventStatus"
      },
      "EmailEventType": {
        "type": "string",
        "enum": [
          "测评",
          "笔试",
          "面试",
          "offer",
          "拒信"
        ],
        "title": "EmailEventType"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "PERMIT_REQUIRED",
          "PERMIT_INVALID",
          "STATE_CONFLICT",
          "VALIDATION_ERROR"
        ],
        "title": "ErrorCode"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "统一错误信封（§3 通用约定）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}